Move Function and Macro to respective files Refactor Arguments out
