+++ spsp/macro.py
@@ -1,0 +1,10 @@
+from spsp.function import Function
+
+
+__all__ = [
+    'Macro'
+]
+
+
+class Macro(Function):
+    pass
